refactor(cli): move session recent/prune into cli_session.py
Loose end from the session-group extraction: `session recent` and
`session prune` were defined in cli.py decorated with @session_group.command
(attaching to the imported group object — functional but split across files).
Consolidate them into cli_session.py alongside checkpoint/autosave/
refresh-summary so the whole group lives in one module.

Verification: ruff clean, mypy clean (87 files), session group exposes all
5 subcommands (autosave, checkpoint, prune, recent, refresh-summary),
99 cli/session tests pass.
